Address review notes on the slope NaN early-out PR (#3739)
Note in SlopeNaN that its dask timing forces the compute while
Slope.time_slope does not, so the two classes' dask numbers are not
comparable. Build the speckle mask from a float32 draw to halve the
setup allocation at nx=10000. State in the footprint test comment that
the tests pass on the old kernel too, since they pin behaviour rather
than reproduce a regression. Drop the stale "(unchanged)" from the
planar section banner.

+++ benchmarks/benchmarks/slope.py
@@ -29,10 +29,13 @@ class SlopeNaN(Benchmarking):
         ny = nx // 2
         agg = get_xr_dataarray((ny, nx), type, include_nan=True)
         rng = np.random.default_rng(71942)
-        speckle = rng.random((ny, nx)) < 0.3
+        speckle = rng.random((ny, nx), dtype=np.float32) < 0.3
         self.xr = agg.where(~speckle)
 
     def time_slope_nan(self, nx, type):
+        # Force the compute so the dask case times the kernel rather than
+        # graph construction. Slope.time_slope does not, so the dask numbers
+        # of the two classes are not directly comparable.
         result = self.func(self.xr)
         if hasattr(result.data, "compute"):
             result.data.compute()
